Cleanup and integration with abondis work

+++ jeto/services/api_keys.py
@@ -1,8 +1,8 @@
 # -=- encoding: utf-8 -=-
-from flask import request
+from flask import request, abort
 
-from flask.ext.login import current_user
-from flask.ext.restful import fields, marshal_with
+from flask_login import current_user
+from flask_restful import fields, marshal_with
 
 from jeto import db
 
@@ -10,10 +10,6 @@ from jeto.models.api import APIKeys
 from jeto.services import RestrictedResource  # , adminAuthenticate
 from jeto.services.users import user_fields
 from uuid import uuid4
-
-
-json_headers = {'Content-Type': 'application/json',
-                'Accept': 'application/json'}
 
 api_key_fields = {
     'id': fields.Integer,
@@ -23,27 +19,46 @@ api_key_fields = {
 
 
 class APIKeyApi(RestrictedResource):
-    @marshal_with(api_key_fields, envelope='keys')
-    def get(self, id=None):
+    @marshal_with(api_key_fields)
+    def get(self, userId=None, id=None):
         """Retrieve a list of API keys"""
         if id is not None:
-            return APIKeys.query.get(id)
-        marsh = APIKeys.query.all()
-        return marsh
+            key = APIKeys.query.get(id)
+            if key.user == current_user or current_user.is_admin():
+                return key
+            else:
+                abort(403)
 
+        if userId == current_user.id or current_user.is_admin():
+            marsh = APIKeys.query.filter(APIKeys.user_id == userId).all()
+            return marsh
+        else:
+            abort(403)
+
+    @marshal_with(api_key_fields)
     def post(self):
         query = request.args
         comment = query.get('comment')
+
+        # @TODO : We should not bind the user to current_user, in case the
+        # key was added by an admin to a user.
         user = current_user
         api_key = APIKeys()
+
+        # @TODO : We should make sure the name was not already provided, maybe
+        # we are just changing the comment on this key.
         api_key.name = unicode(uuid4())
         api_key.user = user
-        api_key.name = comment or "Random API Key"
+        api_key.comment = comment or "Random API Key"
         db.session.add(api_key)
         db.session.commit()
+        return api_key
 
-    def delete(self, id):
+    def delete(self, userId, id):
         """delete API Key"""
         key = APIKeys.query.get(id)
-        db.session.delete(key)
-        db.session.commit()
+        if key.user == current_user or current_user.is_admin():
+            db.session.delete(key)
+            db.session.commit()
+        else:
+            abort(403)
